Fix the cookie cutter template for beat
Convert all template variables into cookie cutter namespace and double
braces
