+++ generate/beat/cookiecutter.json
@@ -1,0 +1,7 @@
+{
+  "project_name": "Examplebeat",
+  "github_name": "your-github-name",
+  "beat": "{{ cookiecutter.project_name|lower }}",
+  "beat_path": "github.com/{{ cookiecutter.github_name|lower }}",
+  "full_name": "Firstname Lastname"
+}
