--- generate/beat/{beat}/_meta/beat.yml
@@ -1,7 +1,0 @@
-################### {Beat} Configuration Example #########################
-
-############################# {Beat} ######################################
-
-{beat}:
-  # Defines how often an event is sent to the output
-  period: 1s
