+++ generate/beat/{{cookiecutter.beat}}/.travis.yml
