+++ generate/beat/{{cookiecutter.beat}}/LICENSE
@@ -1,4 +1,4 @@
-Copyright (c) 2016 {full_name}
+Copyright (c) 2016 {{cookiecutter.full_name}}
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
