+++ generate/beat/{{cookiecutter.beat}}/NOTICE
@@ -1,5 +1,5 @@
 {beat}
-Copyright 2017 {full_name}
+Copyright 2017 {{cookiecutter.full_name}}
 
 This product includes software developed by The Apache Software
 Foundation (http://www.apache.org/).
