+++ generate/beat/{{cookiecutter.beat}}/_meta/beat.yml
@@ -1,0 +1,7 @@
+################### {{cookiecutter.beat|capitalize}} Configuration Example #########################
+
+############################# {{cookiecutter.beat|capitalize}} ######################################
+
+{{cookiecutter.beat}}:
+  # Defines how often an event is sent to the output
+  period: 1s
