+++ generate/beat/{{cookiecutter.beat}}/config/config.go
