+++ generate/beat/{{cookiecutter.beat}}/config/config_test.go
