+++ generate/beat/{{cookiecutter.beat}}/main_test.go
@@ -1,6 +1,6 @@
 package main
 
-// This file is mandatory as otherwise the {beat}.test binary is not generated correctly.
+// This file is mandatory as otherwise the {{cookiecutter.beat}}.test binary is not generated correctly.
 
 import (
 	"flag"
