+++ generate/beat/{{cookiecutter.beat}}/tests/system/test_base.py
@@ -1,0 +1,19 @@
+from {{cookiecutter.beat}} import BaseTest
+
+import os
+
+
+class Test(BaseTest):
+
+    def test_base(self):
+        """
+        Basic test with exiting {{cookiecutter.beat|capitalize}} normally
+        """
+        self.render_config_template(
+                path=os.path.abspath(self.working_dir) + "/log/*"
+        )
+
+        {{cookiecutter.beat}}_proc = self.start_beat()
+        self.wait_until( lambda: self.log_contains("{{cookiecutter.beat}} is running"))
+        exit_code = {{cookiecutter.beat}}_proc.kill_and_wait()
+        assert exit_code == 0
